pyproject.toml
[project]
name = "manus-sandbox-env"
version = "2.0.0"
description = "AI agent sandbox with automated testing system for agent validation, API smoke tests, SQLite integrity, and build verification"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "S"]
ignore = ["D100", "D104", "S608", "S603", "S607"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--cov=tools",
    "--cov=scripts",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
markers = [
    "live: tests that require live services (deselect with -m 'not live')",
]
